# Conflicts:
#	.github/workflows/build.yml

+++ src/feeds.c
@@ -27,10 +27,11 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 #include "app.h"
 
 #include <the_Foundation/file.h>
-#include <the_Foundation/mutex.h>
 #include <the_Foundation/hash.h>
-#include <the_Foundation/queue.h>
+#include <the_Foundation/intset.h>
+#include <the_Foundation/mutex.h>
 #include <the_Foundation/path.h>
+#include <the_Foundation/queue.h>
 #include <the_Foundation/regexp.h>
 #include <the_Foundation/stringset.h>
 #include <the_Foundation/thread.h>
@@ -55,12 +56,36 @@ void deinit_FeedEntry(iFeedEntry *d) {
     deinit_String(&d->url);
 }
 
+const iString *url_FeedEntry(const iFeedEntry *d) {
+    const size_t fragPos = indexOf_String(&d->url, '#');
+    if (fragPos != iInvalidPos) {
+        return collect_String(newRange_String((iRangecc){ constBegin_String(&d->url),
+                                                          constBegin_String(&d->url) + fragPos }));
+    }
+    return &d->url;
+}
+
+iBool isUnread_FeedEntry(const iFeedEntry *d) {
+    const size_t fragPos = indexOf_String(&d->url, '#');
+    if (fragPos != iInvalidPos) {
+        /* Check if the entry is newer than the latest visit. */
+        const iTime visTime = urlVisitTime_Visited(visited_App(), url_FeedEntry(d));
+        return cmp_Time(&visTime, &d->posted) < 0;
+    }
+    if (!containsUrl_Visited(visited_App(), &d->url)) {
+        return iTrue;
+    }
+    return iFalse;
+}
+
 /*----------------------------------------------------------------------------------------------*/
 
 struct Impl_FeedJob {
     iString     url;
     uint32_t    bookmarkId;
     iTime       startTime;
+    iBool       isFirstUpdate; /* hasn't been checked ever before */
+    iBool       checkHeadings;
     iGmRequest *request;
     iPtrArray   results;
 };
@@ -71,6 +96,8 @@ static void init_FeedJob(iFeedJob *d, const iBookmark *bookmark) {
     d->request = NULL;
     init_PtrArray(&d->results);
     iZap(d->startTime);
+    d->isFirstUpdate = iFalse;
+    d->checkHeadings = hasTag_Bookmark(bookmark, "headings");
 }
 
 static void deinit_FeedJob(iFeedJob *d) {
@@ -92,6 +119,7 @@ static const int   updateIntervalSeconds_Feeds_ = 4 * 60 * 60;
 struct Impl_Feeds {
     iMutex *  mtx;
     iString   saveDir;
+    iIntSet   previouslyCheckedFeeds; /* bookmark IDs */
     iTime     lastRefreshedAt;
     int       refreshTimer;
     iThread * worker;
@@ -111,6 +139,21 @@ static void submit_FeedJob_(iFeedJob *d) {
     submit_GmRequest(d->request);
 }
 
+static iBool isSubscribed_(void *context, const iBookmark *bm) {
+    iUnused(context);
+    static iRegExp *pattern_ = NULL;
+    if (!pattern_) {
+        pattern_ = new_RegExp("\\bsubscribed\\b", caseSensitive_RegExpOption);
+    }
+    iRegExpMatch m;
+    init_RegExpMatch(&m);
+    return matchString_RegExp(pattern_, &bm->tags, &m);
+}
+
+static const iPtrArray *listSubscriptions_(void) {
+    return list_Bookmarks(bookmarks_App(), NULL, isSubscribed_, NULL);
+}
+
 static iFeedJob *startNextJob_Feeds_(iFeeds *d) {
     if (isEmpty_PtrArray(&d->jobs)) {
         return NULL;
@@ -125,7 +168,9 @@ static void trimTitle_(iString *title) {
     const char *start = constBegin_String(title);
     iConstForEach(String, i, title) {
         start = i.pos;
-        if (!isSpace_Char(i.value) && !(i.value < 128 && ispunct(i.value))) {
+        if (!isSpace_Char(i.value) &&
+            /* Dashes or punctuation? */
+            !(i.value == 0x2013 || i.value == 0x2014 || (i.value < 128 && ispunct(i.value)))) {
             break;
         }
     }
@@ -170,11 +215,75 @@ static void parseResult_FeedJob_(iFeedJob *d) {
                         .year = year, .month = month, .day = day, .hour = 12 /* noon UTC */ });
                 pushBack_PtrArray(&d->results, entry);
             }
+            if (d->checkHeadings) {
+                init_RegExpMatch(&m);
+                if (startsWith_Rangecc(line, "#")) {
+                    while (*line.start == '#' && line.start < line.end) {
+                        line.start++;
+                    }
+                    trimStart_Rangecc(&line);
+                    iFeedEntry *entry = new_FeedEntry();
+                    entry->posted = now;
+                    if (!d->isFirstUpdate) {
+                        entry->discovered = now;
+                    }
+                    entry->bookmarkId = d->bookmarkId;
+                    iString *title = newRange_String(line);
+                    set_String(&entry->title, title);
+                    set_String(&entry->url, &d->url);
+                    appendChar_String(&entry->url, '#');
+                    append_String(&entry->url, collect_String(urlEncode_String(title)));
+                    delete_String(title);
+                    pushBack_PtrArray(&d->results, entry);
+                }
+            }
         }
         deinit_String(&src);
         iRelease(linkPattern);
         iEndCollect();
     }
+}
+
+static void save_Feeds_(iFeeds *d) {
+    iFile *f = new_File(collect_String(concatCStr_Path(&d->saveDir, feedsFilename_Feeds_)));
+    if (open_File(f, write_FileMode | text_FileMode)) {
+        lock_Mutex(d->mtx);
+        iString *str = new_String();
+        format_String(str, "%llu\n# Feeds\n", integralSeconds_Time(&d->lastRefreshedAt));
+        write_File(f, utf8_String(str));
+        /* Index of feeds for IDs. */ {
+            iConstForEach(PtrArray, i, listSubscriptions_()) {
+                const iBookmark *bm = i.ptr;
+                format_String(str, "%08x %s\n", id_Bookmark(bm), cstr_String(&bm->url));
+                write_File(f, utf8_String(str));
+            }
+        }
+        writeData_File(f, "# Entries\n", 10);
+        iTime now;
+        initCurrent_Time(&now);
+        iConstForEach(Array, i, &d->entries.values) {
+            const iFeedEntry *entry = *(const iFeedEntry **) i.value;
+            if (isValid_Time(&entry->discovered) &&
+                secondsSince_Time(&now, &entry->discovered) > maxAge_Visited) {
+                continue; /* Forget entries discovered long ago. */
+            }
+            format_String(str, "%x\n%llu\n%llu\n%s\n%s\n",
+                          entry->bookmarkId,
+                          integralSeconds_Time(&entry->posted),
+                          integralSeconds_Time(&entry->discovered),
+                          cstr_String(&entry->url),
+                          cstr_String(&entry->title));
+            write_File(f, utf8_String(str));
+        }
+        delete_String(str);
+        close_File(f);
+        unlock_Mutex(d->mtx);
+    }
+    iRelease(f);
+}
+
+static iBool isHeadingEntry_FeedEntry_(const iFeedEntry *d) {
+    return contains_String(&d->url, '#');
 }
 
 static iBool updateEntries_Feeds_(iFeeds *d, iPtrArray *incoming) {
@@ -186,23 +295,25 @@ static iBool updateEntries_Feeds_(iFeeds *d, iPtrArray *incoming) {
         if (locate_SortedArray(&d->entries, &entry, &pos)) {
             iFeedEntry *existing = *(iFeedEntry **) at_SortedArray(&d->entries, pos);
             /* Already known, but update it, maybe the time and label have changed. */
-            iBool changed = iFalse;
-            iDate newDate;
-            iDate oldDate;
-            init_Date(&newDate, &entry->posted);
-            init_Date(&oldDate, &existing->posted);
-            if (!equalCase_String(&existing->title, &entry->title) ||
-                (newDate.year != oldDate.year || newDate.month != oldDate.month ||
-                 newDate.day != oldDate.day)) {
-                changed = iTrue;
-            }
-            set_String(&existing->title, &entry->title);
-            existing->posted = entry->posted;
-            delete_FeedEntry(entry);
-            if (changed) {
-                /* TODO: better to use a new flag for read feed entries? */
-                removeUrl_Visited(visited_App(), &existing->url);
-                gotNew = iTrue;
+            if (!isHeadingEntry_FeedEntry_(existing)) {
+                iBool changed = iFalse;
+                iDate newDate;
+                iDate oldDate;
+                init_Date(&newDate, &entry->posted);
+                init_Date(&oldDate, &existing->posted);
+                if (!equalCase_String(&existing->title, &entry->title) ||
+                    (newDate.year != oldDate.year || newDate.month != oldDate.month ||
+                     newDate.day != oldDate.day)) {
+                    changed = iTrue;
+                }
+                set_String(&existing->title, &entry->title);
+                existing->posted = entry->posted;
+                delete_FeedEntry(entry);
+                if (changed) {
+                    /* TODO: better to use a new flag for read feed entries? */
+                    removeUrl_Visited(visited_App(), &existing->url);
+                    gotNew = iTrue;
+                }
             }
         }
         else {
@@ -252,18 +363,10 @@ static iThreadResult fetch_Feeds_(iThread *thread) {
             break;
         }
     }
-    postCommandf_App("feeds.update.finished arg:%d", gotNew ? 1 : 0);
     initCurrent_Time(&d->lastRefreshedAt);
+    save_Feeds_(d);
+    postCommandf_App("feeds.update.finished arg:%d", gotNew ? 1 : 0);
     return 0;
-}
-
-static iBool isSubscribed_(void *context, const iBookmark *bm) {
-    iUnused(context);
-    return indexOfCStr_String(&bm->tags, "subscribed") != iInvalidPos; /* TODO: RegExp with \b */
-}
-
-static const iPtrArray *listSubscriptions_(void) {
-    return list_Bookmarks(bookmarks_App(), NULL, isSubscribed_, NULL);
 }
 
 static iBool startWorker_Feeds_(iFeeds *d) {
@@ -272,7 +375,14 @@ static iBool startWorker_Feeds_(iFeeds *d) {
     }
     /* Queue up all the subscriptions for the worker. */
     iConstForEach(PtrArray, i, listSubscriptions_()) {
-        iFeedJob* job = new_FeedJob(i.ptr);
+        const iBookmark *bm = i.ptr;
+        iFeedJob *job = new_FeedJob(bm);
+        if (!contains_IntSet(&d->previouslyCheckedFeeds, id_Bookmark(bm))) {
+            job->isFirstUpdate = iTrue;
+            printf("first check of %x: %s\n", id_Bookmark(bm), cstr_String(&bm->title));
+            fflush(stdout);
+            insert_IntSet(&d->previouslyCheckedFeeds, id_Bookmark(bm));
+        }
         pushBack_PtrArray(&d->jobs, job);
     }
     if (!isEmpty_Array(&d->jobs)) {
@@ -304,45 +414,8 @@ static void stopWorker_Feeds_(iFeeds *d) {
 }
 
 static int cmp_FeedEntryPtr_(const void *a, const void *b) {
-    const iFeedEntry * const *elem[2] = { a, b };
+    const iFeedEntry * const *elem[2] = { a, b };    
     return cmpString_String(&(*elem[0])->url, &(*elem[1])->url);
-}
-
-static void save_Feeds_(iFeeds *d) {
-    iFile *f = new_File(collect_String(concatCStr_Path(&d->saveDir, feedsFilename_Feeds_)));
-    if (open_File(f, write_FileMode | text_FileMode)) {
-        lock_Mutex(d->mtx);
-        iString *str = new_String();
-        format_String(str, "%llu\n# Feeds\n", integralSeconds_Time(&d->lastRefreshedAt));
-        write_File(f, utf8_String(str));
-        /* Index of feeds for IDs. */ {
-            iConstForEach(PtrArray, i, listSubscriptions_()) {
-                const iBookmark *bm = i.ptr;
-                format_String(str, "%08x %s\n", id_Bookmark(bm), cstr_String(&bm->url));
-                write_File(f, utf8_String(str));
-            }
-        }
-        writeData_File(f, "# Entries\n", 10);
-        iTime now;
-        initCurrent_Time(&now);
-        iConstForEach(Array, i, &d->entries.values) {
-            const iFeedEntry *entry = *(const iFeedEntry **) i.value;
-            if (secondsSince_Time(&now, &entry->discovered) > maxAge_Visited) {
-                continue; /* Forget entries discovered long ago. */
-            }
-            format_String(str, "%x\n%llu\n%llu\n%s\n%s\n",
-                          entry->bookmarkId,
-                          integralSeconds_Time(&entry->posted),
-                          integralSeconds_Time(&entry->discovered),
-                          cstr_String(&entry->url),
-                          cstr_String(&entry->title));
-            write_File(f, utf8_String(str));
-        }
-        delete_String(str);
-        close_File(f);
-        unlock_Mutex(d->mtx);
-    }
-    iRelease(f);
 }
 
 iDeclareType(FeedHashNode)
@@ -388,6 +461,7 @@ static void load_Feeds_(iFeeds *d) {
                             node->node.key      = id;
                             node->bookmarkId    = bookmarkId;
                             insert_Hash(feeds, &node->node);
+                            insert_IntSet(&d->previouslyCheckedFeeds, bookmarkId);
                         }
                     }
                     break;
@@ -406,8 +480,9 @@ static void load_Feeds_(iFeeds *d) {
                     if (!nextSplit_Rangecc(range_Block(src), "\n", &line)) {
                         goto aborted;
                     }
-                    const unsigned long long discovered = strtoull(line.start, NULL, 10);
-                    if (discovered == 0) {
+                    char *endp = NULL;
+                    const unsigned long long discovered = strtoull(line.start, &endp, 10);
+                    if (endp != line.end) {
                         goto aborted;
                     }
                     if (!nextSplit_Rangecc(range_Block(src), "\n", &line)) {
@@ -454,6 +529,7 @@ void init_Feeds(const char *saveDir) {
     iFeeds *d = &feeds_;
     d->mtx = new_Mutex();
     initCStr_String(&d->saveDir, saveDir);
+    init_IntSet(&d->previouslyCheckedFeeds);
     iZap(d->lastRefreshedAt);
     d->worker = NULL;
     init_PtrArray(&d->jobs);
@@ -474,13 +550,13 @@ void deinit_Feeds(void) {
     stopWorker_Feeds_(d);
     iAssert(isEmpty_PtrArray(&d->jobs));
     deinit_PtrArray(&d->jobs);
-    save_Feeds_(d);
     deinit_String(&d->saveDir);
     delete_Mutex(d->mtx);
     iForEach(Array, i, &d->entries.values) {
         iFeedEntry **entry = i.value;
         delete_FeedEntry(*entry);
     }
+    deinit_IntSet(&d->previouslyCheckedFeeds);
     deinit_SortedArray(&d->entries);
 }
 
@@ -552,6 +628,9 @@ const iString *entryListPage_Feeds(void) {
     iZap(on);
     iConstForEach(PtrArray, i, listEntries_Feeds()) {
         const iFeedEntry *entry = i.ptr;
+        if (isHidden_FeedEntry(entry)) {
+            continue; /* A hidden entry. */
+        }
         iDate entryDate;
         init_Date(&entryDate, &entry->posted);
         if (on.year != entryDate.year || on.month != entryDate.month || on.day != entryDate.day) {
